requirements.lock.txt
accelerate==1.6.0
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cramjam==2.9.1
exceptiongroup==1.2.2
fastparquet==2024.11.0
filelock==3.18.0
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pdf2image==1.17.0
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
psutil==7.0.0
PyPDF2==3.0.1
pytesseract==0.3.13
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
ruff==0.11.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.0.2
six==1.17.0
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.2.2
tqdm==4.67.1
transformers==4.51.1
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
